lacros-au: Use lacros common options
Now that lacrosfixt.NewConfig* provides the common chrome options
to enable Lacros, switch to using them for lacros autoupdate tests.

Also, it adds a new 'NotSelected' enum to lacros.Selection since
this test is to verify the selection logic in ash-chrome. Thus,
this doesn't need to be run in the Browser CI with the runtime var
lacrosDeployedBinary.

BUG=chromium:1315088
TEST=# Start TLW + TLS locally
TEST=tast run \
TEST=  -var=provision.TLSAddr=127.0.0.1:7152 ${DUT} \
TEST=  lacros.UpdateRootfsToStateful \
TEST=  lacros.Update*

Change-Id: Id1ff8b0aafb6da42049d97843e26452a54080d30