tast-tests: Add wifi.{CheckIntelFWDump, UdevRename} to wifi_matfunc
These 2 tests are running as part of bvt-tast-cq but not in
wifi_matfunc. That makes it much easier to have an at-a-glance view
of the health of our tests by looking at wifi_matfunc results. It's
also better for our partners who are only required to run wifi_matfunc.

The "group:wificell" is required to run wifi_matfunc as pointed by
control file wificell-func-stable. So this CL will make above 2
tests require wificell fixtures.

BUG=b:226981380
TEST=run test_that --board=octopus $DUT suite:wifi_matfunc and then \
     look into logs to confirm that these 2 tests are added

Cq-Depend: chromium:3603048
Change-Id: Ief1480c070210d3fd675245b73740770a65c89c8