tast-tests: ip show link uses default format for more compatability
Prior to this change, "ip show link" was called with the "-br" flag
in order to adjust the output format to be a more brief format
which was more simple to parse. Unfortunately not all
implementations of ip support the "-br" flag, most notably the
busybox version of ip used in OpenWrt. This change updates the ip
runner to use the default format instead and the output parsing
has been updated accordingly.

BUG=b:211517577
TEST=tast unit tests passed

Change-Id: I95e83edda288eec7dfefe925ba725f70d2b27d48