tast-tests: Increase app install wait timeout in ARCProvisioning
Test runs show that 4 minutes are not enough for apps to install
because Play Store sometimes runs extra jobs before apps are installed
and so this timeout is not strictly a function of no. of apps being
installed.

Changing the timeout to 8 minutes for now.

BUG=b:186274474
BUG=b:228355858
TEST=tast run DUT enterprise.ARCProvisioning

Change-Id: I864eefe7f1e2e41b2babb36888bbb09db39ed92f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3606547
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Auto-Submit: Muhammad Hasan Khan <mhasank@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>