network: Add VPNUI test
This patch adds the end-to-end VPN test which simulate user inputs to
create, connect, disconnect, and remove a VPN service in the Chrome.
This patch only adds the test for L2TP/IPsec with PSK, and more tests
will be added after this test is stablized.

Also refactors vpn/connection.go to make it able to prepare the VPN
environment without installing the Service into shill.

BUG=b:224875028
TEST=tast run $DUT network.VPNUI; checked that the VPN service will be
  disconnected and removed if a failure happens after connected
TEST=tast run $DUT network.VPNConnect.ikev2*

Change-Id: Ic84aa698e0906c23d1ff37b06f11f2e9eb2ffc6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3607850
Reviewed-by: Kamil Szarek <kamilszarek@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>