policy: Fix VirtualMachinesAllowed test
The regression to this test was introduced in crrev.com/c/3565380. The |OpenLinuxSubpage| function no longer works as before, in that it no longer clicks on the "Turn on" button for opening the Linux installer. On the other hand |OpenLinuxInstaller| opens the installer and then also clicks on the "Next" button, which this test does not need.

Therefore rename |OpenLinuxInstaller| to |OpenLinuxInstallerAndClickNext| and add a new function |OpenLinuxInstaller| which just opens the installer without clicking "Next". Finally use the new |OpenLinuxInstaller| function in the policy.VirtualMachinesAllowed test.

BUG=b:229183763
TEST=tast run <DUT> policy.VirtualMachinesAllowed

Change-Id: I66004ba43bf45887d99fcfb18938d5a72217d908
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3607932
Commit-Queue: Jana Grill <janagrill@google.com>
Tested-by: Jana Grill <janagrill@google.com>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>