printer: add RejectDocumentFormat test
This test checks that lpadmin correctly handles the case where printers
reject Get-Printer-Attributes requests that contain the document-format
attribute. Older printers sometimes demonstrate this behavior.

This first version of the test exercises the code in lpadmin that
retries the failed Get-Printer-Attributes request without the
document-format attribute, but it does not finish configuring the
printer.

BUG=b:175631323
TEST=tast run --build $DUT_IP printer.RejectDocumentFormat

Cq-Depend: chromium:3607067
Change-Id: Iefa28f9d7816aa1d6a14e4423b909d9a9136233f