autoupdate: Add cross version chrome login test
We already have a vault compat cross version test, but we want another
test that is more "end-to-end". That is, a test that logs in Chrome for
the first time at milestone N, updates to milestone M (>N), then logs in
again. This also serves as the base test for future cross version tests
we want to add that build upon Chrome (like WebAuthn cross version
test). When those tests failed, we can identify better if it's the
target behavior failing across version, or is it login itself that is
failing.

BUG=b:230070328
TEST=run autoupdate.NToMChromeLogin with the method written in
README.md.

Change-Id: I0750da54ad1d15c2cf47b44816815331a52f895f