settings: Improve CheckCertificateVisibleInSystemSettings
Some tests that are using the
CheckCertificateVisibleInSystemSettings method fail and
screenshots show that "User certificate" is not on the
screen (and presumably not clickable).
E.g. see network.CertSettingsPage on kukui.

Add a WaitUntilExists call to wait for the window to
update itself after selecting EAP-TLS.
Add MakeVisible to scroll the window on devices where
the entire settings page doesn't fit on the screen.

BUG=b:200919806
TEST=network.CertSettingsPage

Change-Id: I25354b0e30d9f2ea5e2ec870570cc61e31c327e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3610031
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Michael Ershov <miersh@google.com>
Commit-Queue: Michael Ershov <miersh@google.com>