settings: Improve CheckCertificateVisibleInSystemSettings
Some tests that are using the
CheckCertificateVisibleInSystemSettings method fail and
screenshots show that "User certificate" is not on the
screen (and presumably not clickable).
E.g. see network.CertSettingsPage on kukui.

Add a WaitUntilExists call to wait for the window to
update itself after selecting EAP-TLS.
Add MakeVisible to scroll the window on devices where
the entire settings page doesn't fit on the screen.

BUG=b:200919806
TEST=network.CertSettingsPage

Change-Id: I25354b0e30d9f2ea5e2ec870570cc61e31c327e4