chrome: Add optional OAuth client ID for auth
This change adds the `identity` permission to the test extension and an
optional `oauth2` section. Tests can optionally specify an OAuth client
ID with `chrome.TestExtOAuthClientID("...") and the `oauth2` section
will be populated with the provided ID.

This change allows tests to use the `chrome.identity.getAuthToken` API
to fetch OAuth access tokens for use with Google APIs.

BUG=chromium:1299712
TEST=tast run <DUT> filemanager.DrivefsDssOffline
TEST=tast run <DUT> filemanager.DrivefsGoogleDoc

Change-Id: Iea9c8330334b81316f5dd6d9bee3c375ac49d778