drivefs: Fetch tokens from extension for auth
This change creates a new `oauth2.TokenSource` which is capable of
fetching OAuth access tokens from the test extension and uses it in the
DriveFS tests.

I had to refactor some of how the Drive API was consumed to make it
easier to use the `TokenSource`. The service is now created once with
the `TokenSource` and the `Context` on the request is overridden at
call time.

BUG=chromium:1299712
TEST=tast run <DUT> filemanager.DrivefsDssOffline
TEST=tast run <DUT> filemanager.DrivefsGoogleDoc

Cq-Depend: chrome-internal:4749158
Change-Id: If87ed3702e281136383ae4258d05f13ff205380c