scanapp: Parse scanner name from USB descriptors
The expected scanner name is currently hardcoded, which has been fine
because all the tests load the same set of USB descriptors.  We want to
add a new test that uses a different descriptor, so the printer name
will be different.  Since the name is fully derived from the USB string
descriptors and we're already parsing the file to get VID and PID out,
switch to getting the printable name out of the file as well.

BUG=b:199521284
TEST=tast run '("dep:virtual_usb_printer")'

Change-Id: I633cfe4e810c9688fae76a98cc944d58f1993a4e