faft: When booting to dev mode, press ctrl-D+enter
If we press Ctrl-D too soon, before the firmware menu is ready for it,
then we'll get stuck. By instead pressing Ctrl-D and ENTER repeatedly
then we can boot more reliablbly. This also removes the need for the
PressEnterAtToNorm, since we'll be pressing ENTER in all cases.

Also change to using a growing delay, like the dev mode USB code does.

BUG=b:229637316
TEST=Ran firmware.FwmpDevDisableBoot firmware.BootMode.normal_dev on
jinlon and collis

Change-Id: Ic097094483eb0471ef68e1fbaa8b87b75cc97d59