tast-tests: Add test case for drag-caption-to-snap
This CL adds new test case for drag-caption-to-snap to
arc.SplitView.clamshell_mode*.
We don’t have to test it in tablet mode because there’s no caption
in tablet mode.
This CL defines some new functions in the helper class because we’ll
reuse them in arc.CompatSnap in the follow-up CLs.

BUG=b:217655499
TEST=tast run DUT arc.SplitView.clamshell_mode*

Change-Id: I7854d88d9b2a3a75d0d8cb5c19aeadd5970b0a88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3610465
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>