network: VPNMojoConf verifies shill properties
Currently VPNMojoConf only verifies if a VPN service is created
successfully in shill. This check is weak since it cannot reveal issues
like something goes wrong in the translation process in Chrome, since we
do not do a strict check there. This patch adds the code to check if
properties are written properly to shill. Since it is done by reading
the service properties exposed by D-Bus API, some credential properties
cannot be checked.

BUG=b:216386693
TEST=tast run $DUT network.VPNMojoConf

Change-Id: I8f574cf8be581cc7c8fa1cecc00168a38be4c755