tast-test: Fix policy.UserFeedbackAllowed flakiness
The lacros test is flaky because in a few cases it is not able to find
the Report option. There is a very short timeout added to
WaitUntilExists, which is unnecessary and seems to be causing the test
to fail when it takes longer time to the node to appear.

This timeout is still required to make sure a node disappears, and
seems to be enough.

BUG=chromium:1319330
TEST=tast run DUT policy.UserFeedbackAllowed
Reproduce current flakiness after a few executions. Run test 50 times
with the fix and observe no related failures. Flakiness seems to be
fixed.

Change-Id: Ia9df151509fa61814d3586f46429722c0440f83e