[tast-test]: Remove items from continue section
Test to verify the behaviour of the launcher Continue Section as items
are getting removed by the user.

Makes sure that the Feedback Dialog apprears until the user fills the
feedback form once.

Also, verifies that the section is removed when the minimum number of
tasts required for the section to show is not met (We need 3 tasks for
clamshell mode and 2 for tablet mode)

BUG=b:229790200
TEST=tast run launcher.ShowContinueSection* \
     launcher.RemoveContinueSectionTask*

Change-Id: I1f9b8e641db6108fc71e60a1eaf60ec876c36117