graphics: Support VKMS on VMs in IGT tests
This change allows us to run VKMS-based IGT tests so we can compare
support of features in VKMS to other platforms.

VKMS is a virtual and headless display driver that runs as in
installable module. This module is available on our VMs images now, and
this change updates the IGT fixture to add and remove the mod during
setup and teardown.

Not all tests run, and some had to be disabled specifically. A detailed
readout of a recent run can be found in this VKMS doc:
https://docs.google.com/document/d/16YdSsZzoOPEZN-kNC8eK4mFHOlnZytbkhEncWDu9kpE/edit#bookmark=id.yi3punjwpxra

BUG=b:229403705
TEST=ran locally

Cq-Depend: chromium:3591357
Change-Id: I53410cb1ec1e7245753f6d5be1775f5bb60cabdb