Use GAIA login on lacros StartupPerf
This CL ports lacros StartupPerf to use GAIA login flow instead.

It incurs a longer login time than the testing "fake" account, but this
shouldn't be a problem when analyzing the results because here we're
interested in the relative measurements between Lacros and Ash-Chrome.

Also in this CL, there's an important fix in rootfs_primary variant that
was misusing LacrosSidebySide mode instead the correct one
LacrosPrimary.

BUG=chromium:1270977
TEST=tast run device lacros.StartupPerf*

Change-Id: Ia0362125d4a7934f518f944dfe28bfa383a1b8f1