tast-tests: repeat channel scan if no probe request is captured in the pcap file
Probe requests from the DUT may not be captured in the pcap file
as the AP keeps sending beacon frames. To prevent this from failing
the test, repeat the channel scan until the probe request is captured
in the pcap file, or it reached the max number of retries.

BUG=b:225073589
TEST=tast run -var=router=localhost:2210 -var=pcap=localhost:2220
localhost:2200  wifi.ChannelScanDwellTime
Append "passive" to the args in iw.TimedScan() to create
ErrNoProbeRequest event

Change-Id: Ie8f303ad8a75d71971161bcb799e812b0f0fafd7