graphics: opencl-cts: differentiate triaged and untriaged failure
Some test are expected to pass because the implementation is complete
on clvk/clspv side. But it can still fail. Those failures are track by
child issues under b/227133175.
Thus, a test can now be expected to pass but can have a bug that will
be used to track those unexpected (but triaged) failures.

BUG=b:227133175
TEST=./fast_build.sh -T

Change-Id: I78417f59d7e843a11511295e0cdb4ec2bb0f49c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3616521
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Romaric Jodin <rjodin@chromium.org>
Commit-Queue: Romaric Jodin <rjodin@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>