graphics: opencl-cts: differenciate triaged and untriaged failure
Some test are expected to pass because the implementation is complete
on clvk/clspv side. But it can still fail. Those failures are track by
child issues under b/227133175.
Thus, a test can now be expected to pass but can have a bug that will
be used to track those unexpected (but triaged) failures.

BUG=b:227133175
TEST=./fast_build.sh -T

Change-Id: I78417f59d7e843a11511295e0cdb4ec2bb0f49c7