tast: Blocking third-party cookies via policy
The existing test only checked that the policy changes the settings
regarding third-party cookies, but it didn't actually set a cookie and
observer that it gets blocked.

This CL sets up two servers, the "main" server under 127.0.0.1 and a
third-party server serving from localhost. (This distinction is
fortunately enough to make it count as third-party.)

Third-party cookies can only be set via a secure HTTPS connection.
Therefore, this CL extends the https tast-module to include support
for multiple servers at once and for custom HTTP handler functions
that can be set per-file.

Also, a certificate is added for the 127.0.0.1 IP address, the
existing certificate only works localhost. The same CA-cert is used
to sign both of these certificates, so only on Certificate Authority
needs to be configured in the test.

BUG=b:200988635
BUG=chromium:1186217
TEST=tast run dut policy.BlockThirdPartyCookies
TEST=tast run dut policy.BlockThirdPartyCookies.lacros

Change-Id: Ic9ef74e9b965e66f842f69b1042964f1665269cf