tast-tests: crostini: Skip a flaky check.
For unknown reasons the "Resize Linux Disk" dialogue node in the
accessibility tree doesn't get a name when running on DUTs in the lab.
It consistently does locally, consistently does not in the lab,
regardless of device.
Per investigation in b/228922464 and elsewhere it seems like no one in
any affected team over the past year or two has figured out why so we
continue to work around it by changing our checks whenever they start
breaking.
In this case, it's redundant to check for both the dialogue and the
slider (which, at least for now, is getting the right properties set) so
remove the dialogue check entirely.

BUG=b:228922464
TEST=Run test on a lab DUT. Went from consistently failing to passing

Change-Id: I9effc12b49e7e0e335974aab53364a6c9611eb20