Increase timeout for packet capture notification
Increase the timeout for waiting for packet capture notification to
disappear from 5 seconds to 10 seconds to decrease the flakiness in
max-size option test case. The test is flaky on max-size option because
the notification doesn't disappear, I suspect increasing the timeout may
solve this flakiness. 10 seconds seems like an acceptable timeout for
notification.

BUG=b:230472054
TEST=Run "tast run <DUT> debugd.PacketCapture"

Change-Id: I7f3be2c21def7cafc99828169f5e7660e471bd2b