tast-tests: Wait for Play Store view to refresh
enterprise.ARCProvisioning test fails with Play Store is empty because
we don't wait for the view to refreh after the apps have been installed.

In this change, we are going to continue polling for 60 seconds instead
of ending the poll as soon as we find an empty window.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: I43354151b80fa6534f83eb1897b8a05c4b9a7d8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3622573
Auto-Submit: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Commit-Queue: Oleh Lamzin <lamzin@google.com>