tast-tests: Remove launcher parts of ui.DesktopControl
The feature ProductivityLauncher is currently in Finch trials and is
about to be enabled by default. ui.DesktopControl assumes the old
launcher UI and fails when the new UI is enabled.

It may take several tries to enable the feature by default. For now,
remove the launcher parts of this test. Once the feature is enabled,
we can rewrite this part of the test to work with the new UI.

BUG=chromium:1321838
TEST=tast run <dut> ui.DesktopControl

Change-Id: I64e36daa806329c47aedeadaa855f35e2629ca4f