Extend fixture timeout to 10 minutes for crossdevice tests
Extend the fixture timeout to 10 minutes for crossdevice tests. This is
to prevent fixture failures where the OS settings is stuck showing
"Waiting for verification" for the connected Android device. CryptAuth
has a delay for verifying newly added connected devices, even for
accounts on the allowlist for faster verification. The extended setup
timeout should give an allowance for the delay, so hopefully we should
be able to successfully verify more often.

BUG=b:230401333
TEST=tast run <dut> '(group:"cross-device")'

Change-Id: Ic93eb748969c3f1649d91ae69986d99d378819f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3626351
Tested-by: Josh Nohle <nohle@chromium.org>
Reviewed-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Josh Nohle <nohle@chromium.org>
Auto-Submit: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>