tast-tests: Handle binary XML in enterprise.ARCProvisioning.vm
rvc-arc has been switched to use binary XML and it is preventing this
test from reading the packages XML file in order to determine if the
restrictions have been applied to the packages.

In this change, we'll use abx2xml utility of Android to convert the
binary XMl file to plain-text before we read it.

BUG=b:231343274
BUG=b:223867057
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: If02fcb950f7b8b4701c64d2219c2ac5ce2ea8419