tast-tests: turn on TerminalSSH flag
Turn on TerminalSSH flag in advance of chrome uprev where the flag will
be turned on by default.  With TerminalSSH flag on, the Terminal app
will now have an extra Home tab when it launches, and we need to ensure
closing the app closes all tabs.

BUG=b:223076712
TEST=tast run DUT '("group:mainline" && !informational && "name:crostini.*")'

Change-Id: Iec188cf523b856b2ded958742584023c0c34ef6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3625936
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>