policy: Fix test for PrintingBackgroundGraphicsDefault on Ash
This fixes a bug introduced in https://crrev.com/c/3562091, which broke
this test on Ash. This is because `SetUp` will not open a window on Ash,
only on Lacros. To fix the test, we switched to `SetUpWithURL`, which
works consistently on both Ash and Lacros.

BUG=b:229517383
TEST=tast -verbose run $IP policy.PrintingBackgroundGraphicsDefault*

Change-Id: Ie4040e7b2efe36d3f2c439b25060ad04246789b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3627354
Tested-by: Christian Flach <cmfcmf@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>