policy: Fix test for PrintingBackgroundGraphicsDefault on Ash
This fixes a bug introduced in https://crrev.com/c/3562091, which broke
this test on Ash. It looks like due to changes of how Lacros is
launched, it is no longer possible to open a new tab via `SetUpWithURL`
that works on both Ash and Lacros.

What seems to work consistently, though, is to open the tab via
`NewConn`.

BUG=b:229517383
TEST=tast -verbose run $IP policy.PrintingBackgroundGraphicsDefault*

Change-Id: Ie4040e7b2efe36d3f2c439b25060ad04246789b5