[tast-tests] Verify shelf behavior when changing the shelf alignment
This CL checks UI behaviors when changing the shelf alignment. This CL
covers all of shelf alignments, including ShelfAlignmentBottom,
ShelfAlignmentLeft and ShelfAlignmentRight. For each alignment, this CL
checks the following things:

(1) The bounds of the home button
(2) The bounds of the time view
(3) Shelf app alignment
(4) Pin additional icons and verify the shelf app alignment again

This test uses the chrome browser icon so its LacrosStatus is
testing.LacrosVariantNeeded.

BUG=b:200987400
TEST=shelf.VerifyShelfAlignment

Change-Id: I152e1cc0918bf45840a8dc458bca26b82ec87a81