Promote ChromeCrashLoopV2.crashpad_mock_consent and remove ChromeCrashLoop
While ChromeCrashLoopV2.crashpad_mock_consent has a few flakes due to
debugd issues, it's significantly less flaky than
ChromeCrashLoop.crashpad_mock_consent. (Compare
https://tastboard.corp.google.com/test/ui.ChromeCrashLoopV2.crashpad_mock_consent
to
https://tastboard.corp.google.com/test/ui.ChromeCrashLoop.crashpad_mock_consent)
Given that, promote V2 to critical so we can get rid of V1; it's an
overall win for non-flakiness even if imperfect.

Once that's done, V2 is up to the same level as V1 and we can finally
remove the problematic V1 test.

BUG=b:202795944
Fixes: b:202795944
TEST=CQ

Change-Id: I5cd0fbcfddda6cb7db61b173a43404315a49b766