nearbyshare: retry discovery on the sender side in CB->CB tests
Temporarily add retries for the sender attempting to find the receiver.
Certain pairs of devices in the lab are failing much more than others at
this step and we suspect it is a bluetooth platform issue. To get more
insight into the root cause, this CL will toggle bluetooth off and
back on again and retry sending if the sender initially fails to find
the receiver.

Any failures that happen during the retries will have error strings that
start with "(Discovery re-attempt)" so we can easily filter Stainless
results. We will count a test as failing even if the second attempt
succeeds, but the error message will make it clear that's the case.

Once the bug is resolved, these changes should be reverted.

BUG=b:228377059
TEST=tast run -companiondut=cd1:<dut2> <dut1>
'("group:nearby-share-remote")'

Change-Id: I62fb3d355976c63ad8086e572dde9bc976120e0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3630864
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jonathan Fan <jonfan@chromium.org>