tast-tests: Use a different timeout for stable and unstable variant
enterprise.ARCProvisioning does retries in stable variant and does not
do retries in the unstable variant therefore the timeout for both the
tests have to be different.

In this change, we're increasing the timeout for the stable variant to
15 minutes because sometimes Play Store hygiene job can take 10 minutes
to complete so we need more time for the app install to finish.

For the unstable variant, we're reducing the time to 10 minutes because
that is what we expect the ideal run to be like.

We're also removing the timeout for app install specifically and relying
on the test timeout to wait for apps so that it gets the maximum amount
of time left from the over-all test run.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: I57a0c640e217af820e66b5e64c3c41890bed61d6