network: Fix VPNConnect.wireguard test
This patch fixes a issued introduced in https://crrev.com/c/3607850
which causes network.VPNConnect.wireguard_generate_key failed. The
reason is that after the previous patch, createProperties() should be
called before configureService(), which was not added in
generateWireGuardKey(). This patch also refactors createProperties() a
little bit to make it setting |c.Properties| directly instead of
returning the value.

BUG=b:231517021
TEST=tast run $DUT network.VPNConnect.*
TEST=tast run $DUT network.VPNUI

Change-Id: I5b3366647ddab76289209920d545ce8c930a95e0