tast-tests: Add launcher perf test to ui.DesktopControl
Add tests of the new ProductivityLauncher open and close animation
performance to ui.DesktopControl. We had to remove launcher perf
tests earlier because ash-chrome is moving from the legacy peeking
launcher to the new bubble launcher, but we were not sure when the
chrome uprev with the flag flip would happen.

This CL must wait to land until we're sure the flag flip has stuck:
https://chromium-review.googlesource.com/c/chromium/src/+/3630505

BUG=chromium:1322542
TEST=tast run <dut> ui.DesktopControl

Change-Id: Idbeb18bce40b8417a769ecae58df10cbfa5731cc