tast-tests: Delay validation for app install and uninstall
Play Store takes some time to load the UI so do not check for apps
immediately and it does not refresh the view when the app installs so we
need to close and re-open the window.

App uninstall restriction takes time to propagate so instead of
validating it immediately after the package restrictions are written to
the file, we should wait until Play Store UI is updated before attempting
to uninstall the package.

Both of these steps eliminates the flakiness of this test and makes it
run successfully multiple times.

BUG=b:186274474
BUG=b:231751280
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: Ic1f473a2a2905005da2159e963938eb8441980a5