crostini: Include vim tarball in test data, increase timeout
Push the vim tarball as test data, rather than downloading it from
GitHub. Cloning the repo is a significant time sink, pulls in
unnecessary data and is not hermetic, so this should improve the pass
rate of this test even within the previous timeout.

Some slower devices take up to 16 minutes to complete the test, so
increase the timeout as well.

BUG=b:231512028
TEST=tast run <DUT> crostini.VimCompile

Change-Id: I3b5e35b21ae903df14ff6ab2d8b78efce221645d