Cancel context after creating meeting in MeetMultiTaskingCUJ
MeetMultiTaskingCUJ has a context sctx that is used only when calling
bc.CreateConference and may as well be canceled immediately after that.
This reverts advice I gave on CL:3539802, where I suggested to limit the
scope of sctx using simply {}. Calling an anonymous function is slightly
more complicated, but the advantage is that the code "defer cancel()"
defers cancel() only until the end of that anonymous function, whereas
if you just use {} then cancel() gets deferred to the end of the test.

According to [1], in many cases, ctxutil.Shorten starts a new goroutine
associated with the new context, and it is released only on cancellation
or expiration of the context. So I guess that keeping the context much
longer than necessary just might be a bit of a performance issue here.

[1] https://chromium.googlesource.com/chromiumos/platform/tast/+/HEAD/docs/code_review_comments.md#context-cancellation

BUG=b:222110911
TEST=tast run <DUT> ui.MeetMultiTaskingCUJ*

Change-Id: I59635ae4a87b26096a34b31e64ec85b034ffd0a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3635249
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>