[tast-tests] Verify rearranging ARC++ app shelf icon
Verify that shelf app icon arrangement works for ARC++ app icons.

This CL also makes some changes on an Android test app (fully owned by
launcher/shelf team) so that an app window shows when the app is
activated.

The related CL that verifies the shelf icon rearrangement on the File
app and the Chrome app: https://crrev.com/c/2999318

BUG=b:190654715
TEST=shelf.AppRearrangement.rearrange_android_app_android*

Change-Id: I7f6abfd456757a5ac09490f6945614cb0176d689