Try to wait for video in ARC app in WindowArrangementCUJ
This CL is motivated by some slightly problematic behavior of the ARC
app in the WindowArrangementCUJ tests. Sometimes it shows the message
"Can't play this video." for five seconds and then starts playing the
video. Sometimes it just shows the message "Can't play this video."
until the end of the test. Now, the test will wait for the message
"Can't play this video." to disappear.

Unfortunately this CL may not be an ideal solution. I have seen times
when the ARC app appears just black for five seconds and then starts
playing the video. I just am not sure if there is a simple way that the
test could detect that scenario.

BUG=b:226608497
TEST=tast run <DUT> ui.WindowArrangementCUJ.*

Change-Id: Ia843c20ee447f2b173a2db9254078906d53e0a76