network: VPNUI: Change target node after forgetting service
After clicking the "Forget" button on a VPN service, it will be
navigated to "chrome://os-settings/networks?type=VPN" page. Currently we
check if there is a "Built-in VPN" text on the page, but we observed
some flakiness since the contents on this page might be empty (see
Comment #10 on the bug page). This patch changes to use the "VPN" header
node instead: this node is not shown on the previous page but on both of
the two versions on the VPN list page.

BUG=b:224875028
TEST=tast run $DUT network.VPNUI

Change-Id: I45500c85cca4482367b8fcffa05a55cb9caf2918