autoupdate: Add NToMWebauthnLogin test
Add a cross version test that logs in Chrome and creates a WebAuthn
credential in webauthn.io, rollback to ToT, logs in again, and go to
webauthn.io to verify that the same credential can be used to login.

This helps us ensure that the changes in how u2fd processes the key
handles and how the webauthn secret is derived are maintained correctly
throughout version updates for backward compatibility.

BUG=b:230070328
TEST=autoupdate.NToMWebauthnLogin test

Disallow-Recycled-Builds: test-failures
Change-Id: If1a9bcfca1735344e4845db3673e6e634afd8697
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3638032
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>