Tast: fix remote faillog copy
This is a followup to crrev.com/c/3587699. That CL changed the remote
faillog copy so that the remote directory was copied to a local path
with a timestamp, e.g. "faillog/20220510-155500.867568". However,
linuxssh.GetFile does not create parent directories, so that copy only
works if the "faillog" parent directory already exists.

To ensure the copy works in all cases, create the parent destination
directory if it doesn't already exist.

BUG=b:229284804
TEST=Call `faillog.DumpUITree` from a remote service in a remote test that is failing. The remote test must not create a faillog directory.
TEST=Inspect the test results under `/tmp/tast/results/latest/tests/<name>/`
TEST=Verify the remote faillogs have been copied to a new directory, e.g. "faillog/20220510-155500.867568"

Change-Id: I9f90a45f5be8f4c316b5b023532234bdbe982580