scanapp: Optimize scanner idle check
Currently, we run `lorgnette_cli list` each time we check if a scanner
is idle. With this change, we instead run it a single time to get the
scanner's URI at the beginning of the test and use that URI each time we
check if the scanner is idle.

BUG=b:227215718
TEST=tast run --build $DUT_IP scanapp.Hardware.*

Change-Id: I06019fde4edffb22d2902ffb0b1ebe8dac89c42c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3640430
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>