arc: Add arc.ExternalStorageDisabled
This CL adds a tast test arc.ExternalStorageDisabled to ensure that
MyFiles is mounted on ARC on a enterprise-managed device even when the
enterprise policy ExternalStorageDisabled is set to true.
The bug associated to this test (b/217531658) is fixed in
crrev.com/c/3650484. Without this patch, arc.ExternalStorageDisabled.vm
randomly fails or passes, while arc.ExternalStorageDisabled always
passes with or without the patch.

BUG=b:217531658
TEST=tast run <DUT> arc.ExternalStorageDisabled
TEST=tast run <DUT> arc.ExternalStorageDisabled.vm
TEST=tast run <DUT> arc.OptinManaged
TEST=tast run <DUT> arc.OptinManaged.vm

Cq-Depend: chromium:3650484
Change-Id: I89fa2811ffcf5882aae66f2af2dd10dd2b5ea5cc