arc: Add arc.ExternalStorageDisabled
This CL adds a tast test arc.ExternalStorageDisabled to ensure that
the enterprise policy ExternalStorageDisabled is correctly applied for
ARC. In particular, it checks the following two points when the policy
is set to true:
1. MyFiles is mounted on ARC.
2. removable media is not mounted on ARC.
The bug associated to the first part (b/217531658) was fixed in
crrev.com/c/3650484. Also, crrev.com/c/3655844 added an additional
policy check in ArcVolumeMounterBridge in preparation for the second
part.

BUG=b:217531658, b:233308595
TEST=tast run <DUT> arc.EnableExternalStorage*
TEST=tast run <DUT> arc.ExternalStorageDisabled*
TEST=tast run <DUT> arc.OptinManaged*
TEST=tast run <DUT> arc.RemovableMedia*
TEST=Check that `tast run <DUT> arc.ExternalStorageDisabled.vm` fails
     with crrev.com/c/3650484 reverted

Change-Id: I89fa2811ffcf5882aae66f2af2dd10dd2b5ea5cc