tast-tests: Update WiFi6e tast test names/descriptions
The wifi6e tast test names didn't have the 80211ax prefix which caused
problems when trying to run all the ax tests with the regex
`wifi.AxSimpleConnect.80211ax*`. All tests have this prefix now to be
consistent.

Descriptions are updated to indicate which tests are running on the 6ghz
band.

Descriptions originally mentioned 80211ay for 6e tests which is not what
80211ay is. WiFi 6e still falls within 80211ax.

BUG=b:228602405
TEST=tast run wifi.AxSimpleConnect.80211ax*

Change-Id: Iaf77d859bc74d3760ff3b753b98b991f1abcc49f