health: Initiate bluetooth power status in MonitorBluetoothEvent testing
Initially we assumed that the Bluetooth power on the DUT is on by default. And we found that about 10% of DUTs fail to detect a Bluetooth power off for the first time. The reason may be that these DUTs have their Bluetooth powered off by default.

So we initiate the power status off at first.

BUG=b:232322920
TEST=tast run $DUT_IP health.MonitorBluetoothEvent

Change-Id: Iadb4589a59ac18585c795a362e3337d473c1d72a